requirements.txt
cairosvg==2.7.1
networkx==3.2.1
pypng==0.20220715.0
svgwrite==1.4.3
